# fix(pkg): Avoid duplicate `ls-tree` calls