--- a/src/dune_pkg/rev_store.mli
+++ b/src/dune_pkg/rev_store.mli
@@ -12,7 +12,7 @@ module Remote : sig
     type t
 
     val content : t -> Path.Local.t -> string option Fiber.t
-    val directory_entries : t -> Path.Local.t -> Path.Local.t list
+    val directory_entries : t -> Path.Local.t -> Path.Local.Set.t
     val equal : t -> t -> bool
     val repository_id : t -> Repository_id.t
   end