For a particular revision the output of `ls-tree` will never change, so we can just get the file list when the `At_rev.t` is created.