@Alizter 2023-11-01
Testing the locking of `patdiff` with no networking before and after this patch. 

# Before
```
Benchmark 1: ~/dune/_build/install/default/bin/dune pkg lock --disable-networking
  Time (mean ± σ):     32.887 s ±  0.182 s    [User: 40.304 s, System: 13.059 s]
  Range (min … max):   32.601 s … 33.124 s    10 runs
```
# After
``
Benchmark 1: ~/dune/_build/install/default/bin/dune pkg lock --disable-networking
  Time (mean ± σ):      6.792 s ±  0.069 s    [User: 14.942 s, System: 8.382 s]
  Range (min … max):    6.716 s …  6.923 s    10 runs
```