@rgrinberg 2023-11-01
Thanks. I converted the list to a set. This simplifies the implementation somewhat and also avoids equality having that slow sorting step.